refactor: rename to .ts & refactor indents in controller file
