+++ migration-newip-to-ts/src/components/controller/controller.ts
@@ -1,0 +1,39 @@
+import AppLoader from './appLoader';
+
+class AppController extends AppLoader {
+  getSources(callback) {
+    super.getResp(
+      {
+        endpoint: 'sources',
+      },
+      callback,
+    );
+  }
+
+  getNews(e, callback) {
+    let target = e.target;
+    const newsContainer = e.currentTarget;
+
+    while (target !== newsContainer) {
+      if (target.classList.contains('source__item')) {
+        const sourceId = target.getAttribute('data-source-id');
+        if (newsContainer.getAttribute('data-source') !== sourceId) {
+          newsContainer.setAttribute('data-source', sourceId);
+          super.getResp(
+            {
+              endpoint: 'everything',
+              options: {
+                sources: sourceId,
+              },
+            },
+            callback,
+          );
+        }
+        return;
+      }
+      target = target.parentNode;
+    }
+  }
+}
+
+export default AppController;
